Move .range tests to separate file.
